Added functionality of transportPage

+++ Transport/Application/Services/BicycleService.cs
@@ -2,6 +2,7 @@
 using Entities.Contracts.Repositories;
 using Entities.Contracts.Services;
 using Entities.Dto;
+using Entities.Exceptions;
 using Entities.Models.Transports;
 using System;
 using System.Collections.Generic;
@@ -32,24 +33,39 @@ namespace Application.Services
             return bikeDto;
         }
 
-        public async Task<Bicycle> CreateBicycleAsync(CreateBicycleDto createBikeDto)
+        public async Task<IEnumerable<KeyValueDtoObject>> GetSelectedBicycleById(string id)
+        {
+            return await _repository.FindSelectedTransportByIdAsync(id);
+        }
+
+        public async Task<Bicycle> CreateBicycleAsync(CreateBicycleDto createBikeDto, string uid)
         {
             var bike = _mapper.Map<Bicycle>(createBikeDto);
+
+            bike.UserId = uid;
 
             await _repository.CreateBicycleAsync(bike);
 
             return bike;
         }
 
-        public async Task DeleteBicycleAsync(string id)
+        public async Task DeleteBicycleAsync(string id, string uid)
         {
             var bike = await _repository.GetBicycleByIdAsync(id);
+
+            if (!bike.UserId.Equals(uid))
+                throw new DeletionIsNotAllowedException(uid);
 
             await _repository.DeleteBicycleAsync(bike);
         }
 
-        public async Task UpdateBicycleAsync(string id, BicycleDto updateBikeDto)
+        public async Task UpdateBicycleAsync(string id, BicycleDto updateBikeDto, string uid)
         {
+            var currentBicycle = await _repository.GetBicycleByIdAsync(id);
+
+            if (!currentBicycle.UserId.Equals(uid)) 
+                throw new UpdateIsNotAllowedException(uid);
+
             var bike = _mapper.Map<Bicycle>(updateBikeDto);
 
             bike.Id = id;
